[iOS] Add favicon to passkey creation confirmation dialog
If a credential already exists for the rpId for which a passkey
is about to be created, show that favicon instead of the default
world favicon in the passkey creation confirmation dialog shown
when a user is using multi-profile.

This CL extracts from CredentialListViewController the favicon
fetching code into a new utility file and uses this new utility
function in the MultiProfilePasskeyCreationViewController to
fetch the favicon, if possible. The default world icon is still
used as the default favicon and fetching the "correct" favicon
is a low priority operation.

Bug: 382479915
Change-Id: I2e85d6941a18c2f84faf41fe29f9918594ebdb83
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6425999
Reviewed-by: Tommy Martino <tmartino@chromium.org>
Commit-Queue: Alexis Hétu <sugoi@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1441795}