[AXTreeFixing] Add screenshotter
Add screenshotter helper class which captures a full page, in-memory,
screenshot of the given web contents using Paint Preview.

When a screenshot is requested, we use PaintPreviewBaseService to
capture a paint preview, and then we use the PaintPreviewCompositor API
to create a bitmap from it.

The screenshots can be passed to downstream services via the
AXTreeFixingServicesRouter to process.

Bug: 395134535, 401034576, 406272896
Change-Id: I9e0bcac97d55be645ddf5271ce75bd81326fb5a7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6404535
Reviewed-by: Mark Schillaci <mschillaci@google.com>
Commit-Queue: Jocelyn Tran <jocelyntran@google.com>
Cr-Commit-Position: refs/heads/main@{#1441802}