[PWA] Remove uninstall_requests_, clean up policy manager unit tests
The number of install_requests_ and uninstall_requests_ made to the
ExternallyManagedAppManager is an implementation detail that has been
exposed via the FakeExternallyManagedAppManager, which shouldn't be
available.

This CL removes all traces of uninstall_requests_, and some traces of
the install_requests_ from the web_applications/ code. This is the first
step in making InstallApps() and UninstallApps() a private API.

To do this, the WebAppPolicyManager unit tests have been cleaned up so
that they only rely on the final state, which is whether the app has
been installed or not in the registry and has the correct fields
populated. The only tests that have not been easy to migrate are:

1. ForceInstallAppWithCustomAppIcon -> custom_app_icon logic is only
used for placeholder apps. 2. ForceInstallAppWithCreateDesktopShortcut
-> create_desktop_shortcut logic is only used for placeholder apps it
seems, and doesn't update the OS integrations state at all.

Bug: 407549320
Change-Id: Ib7a9feaaafbc82b585ea17fa12230be300ea5668
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6419396
Reviewed-by: Daniel Murphy <dmurph@chromium.org>
Reviewed-by: Jimmy Gong <jimmyxgong@chromium.org>
Commit-Queue: Dibyajyoti Pal <dibyapal@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1441825}