CrOS Settings: Convert supportedSettingIds to a plain class member
The supportedSettingIds Polymer property defined in DeepLinkingMixin
does not need to be a Polymer property since it isn't used in HTML nor
does it use any Polymer property capabilities. Convert this to a plain
class member instead. In some cases, the supportedSettingIds set is
obsolete and can be removed.

This is done in preparation for enabling the ESLint rule
webui-eslint/polymer-property-class-member.

BYPASS_LARGE_CHANGE_WARNING since this is a trivial change across many
files.

Bug: 406584477
Test: browser_tests
Change-Id: Ic1beb4b21c4d8c9ac3ec46b999929b81d44ce874
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6427981
Commit-Queue: Wes Okuhara <wesokuhara@google.com>
Reviewed-by: Demetrios Papadopoulos <dpapad@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1441886}