WebView Prerender: Ignore X prefixed headers on header match during activation
This CL allows to ignore X prefixed headers on header match during
prerender activation check only on WebView, as WebView can add arbitrary
headers to (prerendering) navigation, and that easily causes header
mismatches.

This change is compatible with the prerender spec, as the spec doesn't
require header comparison in the first place, and this check was
conservatively added just for being on the safe side in the initial
implementation. We plan to expand this relaxation to other platforms.

Change-Id: I7ef1000341b78e0d430ad599be65f537420e124d
Bug: 40244149
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6424220
Reviewed-by: Kouhei Ueno <kouhei@chromium.org>
Commit-Queue: Kouhei Ueno <kouhei@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1441910}