Add ApplicationLocaleStorage in //components
Currently, application locale string is managed in BrowserProcessImpl,
and there are many CrOS specific components that depend on the locale
string. Due to the dependency, those components need to depend on
browser_process even if they only require locale.

This CL moves the application locale string to a dedicated class, and
let GlobalFeatures own the instance. After this change, we can pass the
reference to classes that require locale.

Bug: 403155091
Change-Id: I24fc2247135726e45b97cb0e68d25c3a1fc4ecb1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6306749
Reviewed-by: Erik Chen <erikchen@chromium.org>
Reviewed-by: Ted Choc <tedchoc@chromium.org>
Commit-Queue: Jun Ishiguro <junis@google.com>
Cr-Commit-Position: refs/heads/main@{#1441953}