[InputVizard] Relax check that RenderInputRouterSupport is a root frame
When a touch sequence is successfully transferred to Viz, the Browser
sends state corresponding to the touch sequence for Viz to start it's
input handling pipeline. On Browser side RenderWidgetHostViewAndroid
sends the state and includes it's frame sink id in state, which is then
used to query a corresponding RenderInputRouterSupport on Viz. It's
expected the RIRSupport would not be a child support on Viz side but due
to an unidentified bug it seems to happen.

For now this change relaxes the check to avoid crash in such scenarios.
This would result in dropping the touch sequence on Viz side, we can
track the dropped sequences from
`Android.InputOnViz.Viz.StateProcessingResult` histogram with enum value
as `kFrameSinkIdCorrespondsToChildView`.

Bug: 404741207
Change-Id: Idbc52c08e1627e11d6847982fa796a5fc74f88ef
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6425560
Reviewed-by: Aman Verma <amanvr@google.com>
Commit-Queue: Kartar Singh <kartarsingh@google.com>
Cr-Commit-Position: refs/heads/main@{#1442042}