[ios] Fix fullscreen left in partial state when scroll ends
This change fixes an issue where the fullscreen progress could be left
in a partial state at the end of a scroll event. When the drag ends if
there will be no deceleration then the scroll should be marked as ended
immediately. If there is deceleration, then the scroll event will be
ended when the deceleration ends (as is already implemented).

Fixed: 407868629
Change-Id: Id4e6773cb3ce191058ee9ba6453f4379dd7d27ad
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6426419
Reviewed-by: Gauthier Ambard <gambard@chromium.org>
Commit-Queue: Scott Yoder <scottyoder@google.com>
Cr-Commit-Position: refs/heads/main@{#1442053}