Fix  PermissionPromptSurvey to trigger upon tab closure.
Tab closure causes active prompts to be ignored, which is a supported
prompt action trigger of the permission prompt survey. However, tab
closure causes a tab to be destructed, which causes the WebContents to
have Visibility::HIDDEN. The HaTS service checks that a WebContents is
visible if a survey should be triggered for a particular WebContents,
hence tab closure currently won't ever trigger a HaTS survey.

To support this functionality, we instead launch the survey for the last
active browser when the trigger is run. This requires some HaTS service
plumbing, since the permission prompt survey supports custom invitations
on Clank, which is an experimental feature for which we still need to
make a decision on, and hence should not remove support for at this
point.

Bug: 407953263
Change-Id: I5f5bb48b442b620e6b0faec47f160be9e2af9ec0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6425143
Commit-Queue: Florian Jacky <fjacky@chromium.org>
Reviewed-by: Martin Šrámek <msramek@chromium.org>
Reviewed-by: Alex Ilin <alexilin@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1442057}