[UPMLocalDbSunset] Remove fake GMS Core version from some tests
Setting the fake GMS Core version was originally added because the tests
ended up hitting the empty password store backend when clearing cookies
on devices with a too low GMS version.

The empty store backend doesn't accept write operations, because it
cannot store data. While disabling auto sign-in for an origin (which
happens when deleting cookies) is technically a modifying operation,
there is no harm in allowing it to be called on an empty backend. After
all the backend is already not supporting auto-sign in since it doesn't
store passwords.

https://crrev.com/c/6421881 started allowing this call, so the tests no
longer need to fake the GMS Core version.

Bug:407535752

Change-Id: I7a0d35e20496c5de53a21cc8e85027e0e62ce2b8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6425900
Reviewed-by: Christian Dullweber <dullweber@chromium.org>
Commit-Queue: Ioana Pandele <ioanap@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1442083}