[iOS] Add spec for suggesting password in add password panel
Add spec for suggesting password in add password panel.
If a valid URL is provided and a spec was found for that URL,
the suggested password will be generated based on that spec.
Otherwise, the suggested password will be based on the default spec.

Fixed: 404259053
Change-Id: I6f040ee8ea9ec7ffaa6adb30534ed10b3ad7918f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6361055
Reviewed-by: Maria Kazinova <kazinova@google.com>
Commit-Queue: Prasanaa Vingadassamy <prasanaa@google.com>
Reviewed-by: Tommy Martino <tmartino@chromium.org>
Feels: Maria Kazinova <kazinova@google.com>
Cr-Commit-Position: refs/heads/main@{#1442096}