[Tab Grid Dialog] Fix exception if drawing 0 size bitmap
In rare cases a view for the tab group color dot is not laid out yet
and we get an IllegalArgumentException when creating a bitmap for a
dimension with 0 size. Fix this and set the drawable to null if this
occurs.

Fixed: 408143237
Change-Id: Id4842b0a5c957b181873b26b101fbe0da8a37d78
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6426149
Reviewed-by: Fiaz Muhammad <mfiaz@google.com>
Commit-Queue: Calder Kitagawa <ckitagawa@chromium.org>
Auto-Submit: Calder Kitagawa <ckitagawa@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1442113}