Store ContentsWebViews in a vector in MultiContentsView and update SetWebContents
This should visually be a no-op. Storing ContentsWebViews in a vector
and making the first in the vector always visible will make
MultiContentsView more scalable without really increasing complexity.
Changing SetWebContents from being based on whether it is the active
ContentsWebView to an index is more reliable and allows us to set the
webcontents before setting the active position.

Bug: 408052275
Change-Id: I4514acac7273d8f4ec91f32ba1029ed4a4a4929f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6427883
Commit-Queue: Caroline Rising <corising@chromium.org>
Reviewed-by: Emily Shack <emshack@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1442131}