[NTP Customization] Add metrics
This CL adds metrics to track the number of times the bottom sheet is
opened and toggles are used to disable/enable magic stack modules.

In this CL, the following three histograms are added:
1.NewTabPage.Customization.{ModuleToggleState}
2.NewTabPage.Customization.BottomSheet.Shown
3.NewTabPage.Customization.OpenBottomSheetEntry

Besides, a user action "MobileMenuNtpCustomization" is added to the
ChromeTabbedActivity and records the user clicking the main menu to open
the bottom sheet.

Bug: 376238770
Change-Id: Ieef241f7893accec2733a8e547ebd335a3f0fa70
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6404356
Reviewed-by: Xi Han <hanxi@chromium.org>
Reviewed-by: Calder Kitagawa <ckitagawa@chromium.org>
Commit-Queue: Yanling Hu <yanlinghu@google.com>
Cr-Commit-Position: refs/heads/main@{#1442157}