Close lens controller when '@page' keyword mode is exited
This CL adds notification plumbing for the OmniboxClient when keyword
mode is entered via OmniboxEditModel::EnterKeywordMode or exited via
ClearKeyword. This is used by the ChromeOmniboxClient to close lens
via the lens controller when the '@page' keyword mode is exited.
The UI may or may not be showing, but this stops contextualization.

Bug: 408030457
Change-Id: Id41ef0a6b7381897faad94bfbbe6dc01718a1122
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6428735
Commit-Queue: Orin Jaworski <orinj@chromium.org>
Reviewed-by: Nihar Majmudar <niharm@google.com>
Reviewed-by: Orin Jaworski <orinj@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1442208}