Migrate some HashValue::data() calls to HashValue::span().
And use safe operations with them when possible. The goal is to remove
data() entirely in favor of span(). Some callers may still need to use
data() to call 3P APIs, but requiring span().data() encourages
consumers that can get away with using spans to do so, and makes it
easier to ensure we've migrated consumers where possible.

Bug: 390223051
Change-Id: Iadcda816b9fa05a95883e17925db56e42a2f4260
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6396146
Reviewed-by: Elly FJ <ellyjones@chromium.org>
Commit-Queue: mmenke <mmenke@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1442226}