[B4B] Migrate extensions call to be the MojoHandler
Removes (as much as possible) the dependency of the Side Panel towards
the Chrome Extensions bookmarks API. This allows to be consistent
throughout all calls in the SidePanel. Going through the Handler/Mojo
framework allows to ensure we use the MergedService (extensions API
does not support merge view).

Exceptions:
- Edit dialog update: this dialog will only be used by users that
have B4B disabled - and since it has no parent_id computation (no risk
of misusage of the Side Panel ID), it doesn't need to be adapted since
it will be removed soon.
- Drag and Drop dependency to `chrome.bookmarkManagerPrivate.*` drag
events. This allows to properly get the information from sources
that are not the Side Panel with ease. In a followup CL the parent
target folder of the Drop action will need to be adapted though, since
it needs side panel parent computation.

This CL also fixes (partially - investigation added in a TODO)
selecting nodes that are linked (parent - children) to be deleted.
This is possible when using compact view with the tree view feature.

Bug: b:380806881, 407986687
Change-Id: I2fcf426a5a2e8a7dd0e82718154c2605a6debe0b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6426047
Reviewed-by: Antonio Sartori <antoniosartori@chromium.org>
Commit-Queue: Ryan Sultanem <rsult@google.com>
Reviewed-by: Emily Shack <emshack@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1442233}