Use TabDialogManager in LensPermissionBubbleController
This CL updates LensPermissionBubbleController to use its tab's
TabDialogManager to create a dialog widget, instead of the previously
used `constrained_window::ShowWebModalDialogViews`.

The difference in behaviour is that now the dialog widget will be closed
if the tab detached (e.g. moving the tab to a new browser).

Bug: 408179468
Change-Id: Ic64abe02e4ce3082e50f3e9d5dcee95de10c8d11
Fixed: 408179468
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6428307
Reviewed-by: Allen Bauer <kylixrd@chromium.org>
Reviewed-by: Erik Chen <erikchen@chromium.org>
Commit-Queue: Kaan Alsan <alsan@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1442262}