[a11y] Set link movement method only when necessary.
Set LinkMovementMethod will make the text focusable. As a result, the
view will be focused by keyboard even if the view is not actionable,
i.e. not including a link.

This Cl sets LinkMovementMethod only when the target string includes a
clickableSpan or UrlSpan.

Bug: 403313357, 397463333
Change-Id: I44575a4c14416b856db24e3786dcb01a6c815bec
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6397963
Commit-Queue: Lijin Shen <lazzzis@google.com>
Reviewed-by: Aishwarya Rajesh <aishwaryarj@google.com>
Reviewed-by: Theresa Sullivan <twellington@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1442276}