[NTP Customization] Hide the "Discover feed" option in the main bottom sheet when necessary
This CL adds the logics to hide the "Discover feed" option in the main
bottom sheet if the feed section is completely removed from the new tab
page. In other words, it is removed if the user is in one of the EEA
countries and the searching engine using is not the google one.

The change of the CL is shown in the following documentation:
https://docs.google.com/document/d/1bMRVyyqOd0-cAmyTbL9IhKNU9uSvDdpdCngSzSXIHnk/edit?tab=t.0#bookmark=id.yhcjjq3eiz9t

Bug: 376238770
Change-Id: I8a8d51788952339202334cc7c5764d2e6c423ff9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6397805
Reviewed-by: Jian Li <jianli@chromium.org>
Reviewed-by: Xi Han <hanxi@chromium.org>
Reviewed-by: Calder Kitagawa <ckitagawa@chromium.org>
Commit-Queue: Yanling Hu <yanlinghu@google.com>
Cr-Commit-Position: refs/heads/main@{#1442284}