[blink] Generate SyncToken in advance in RasterRecord()
Currently in RasterRecord(), WillDrawInternal() sets a boolean that,
if true, will lead to the generation of a SyncToken at a later time.
However, in order to apply scoped SyncToken access for raster, we need
to ensure that the SyncToken's generation is localized inside
RasterRecord().

This CL is a step towards that purpose. We now generate the SyncToken
right after the RasterRecordOOP() operation. This CL is not meant to
cause any behavioral change.

Bug: 40286368
Change-Id: I8b8a0a121b5d2caef8722e4939669770775c86f2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6426236
Commit-Queue: Mingjing Zhang <mjzhang@chromium.org>
Reviewed-by: Vasiliy Telezhnikov <vasilyt@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1442304}