[Blink] Invoke OnResourceReturnedByCompositor() only in CanvasResourceSI
This CanvasResourceProvider method is implemented only by
CanvasResourceProviderSharedImage, which creates only
CanvasResourceSharedImage instances. This facilitates changing this
interface to pass a CanvasResourceSharedImage in a followup, which will
then unlock CanvasResourceProviderSharedImage being able to hold
CanvasResourceSharedImage instances rather than CanvasResource
instances.

Note: This is a reland that fixes a use-after-free in the original CL.

Bug: 352263194
Change-Id: I25dd3759eae0bbf83680179875a460a69dc6d384
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6429424
Reviewed-by: Vasiliy Telezhnikov <vasilyt@chromium.org>
Commit-Queue: Colin Blundell <blundell@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1442308}