[//cc] Pass format explicitly to ZeroCopyRasterBufferProvider
Rather than passing RasterCaps into ZeroCopyRBP and having it obtain the
format from there, pass in the format explicitly. Followup work will do
this for all other providers and then have the RasterBufferProvider base
class take in the format and eliminate RasterBufferProvider::GetFormat()
being virtual.

Note: I considered having RasterBufferProvider take in
RasterCapabilities as an alternative, but it's kind of a headache to do
so thanks to this [1].

[1] https://source.chromium.org/chromium/chromium/src/+/main:cc/test/fake_raster_buffer_provider.cc;l=35-38?q=FakeRasterBufferProvid&ss=chromium

Bug: 403372453
Change-Id: Id62043747e0964df9bcd3367426a94c56271a9eb
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6429564
Commit-Queue: Colin Blundell <blundell@chromium.org>
Reviewed-by: Vasiliy Telezhnikov <vasilyt@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1442329}