Create PolicyDecision for enterprise url check results
This patch is part of the effort to implement Enterprise url filtering
in Chrome iOS. The CL processes the results of checking navigated urls
against the Enterprise backend. For blocked urls, a PolicyDecision is
generated for displaying an Enterprise blocking interstitial.
For warning results, a PolicyDecision is generated for displaying a
warning interstitial.

A subsequent CL deals with displaying the Enterprise interstitials.

Fixed: 399377055
Change-Id: I3d2758d286370e2fef9f99c861036b434895d291
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6416525
Reviewed-by: Mark Cogan <marq@chromium.org>
Reviewed-by: Daniel White <danieltwhite@google.com>
Reviewed-by: Dominique Fauteux-Chapleau <domfc@chromium.org>
Commit-Queue: Ernesto Izquierdo Clua <eic@google.com>
Cr-Commit-Position: refs/heads/main@{#1442343}