[AXTreeFixing] Add short delay and re-request for transient state
This CL cleans up some parts of the AXTreeFixing services router to be
more convenient for clients/testing, and to cover an edge-case.

With this CL, we add a retry mechanism for requests for main nodes in
favor of the previously existing queue (which is kept in the code
with a TODO to address it after final UX). There can be short
transient states, like ScreenAI starting up, that result in the need
to retry a request after a short delay if we do not store a queue in
a state.

We update the impl in web_contents to no longer use hard CHECKs, to
cover the case where a page is already fully loaded before a user
turns on accessibility services. In this case, the observers and the
services router can be spun up before the BrowserAccessibilityManager
for each RenderFrame is ready, so rather than doing a CHECK, we
revert to the original implementation of returning an empty AXTree,
and using that as a signal to retry the request after a short delay.

We update some comments and add a request counter member variable to
ensure we don't do many repeated requests, which is a sign of
something wrong that is not a transient state.

AX-Relnotes: N/A
Bug: 395134535, 401050816
Change-Id: I604b0fab822887b689419664e48a573bbedb9d76
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6426603
Reviewed-by: Alex Moshchuk <alexmos@chromium.org>
Reviewed-by: Jocelyn Tran <jocelyntran@google.com>
Commit-Queue: Mark Schillaci <mschillaci@google.com>
Cr-Commit-Position: refs/heads/main@{#1442369}